DogFinder/face-detection/requirements.txt
Flask
flask-cors
flask_sqlalchemy
keras_cv
tensorflow
opencv-python-headless
Pillow
